backend/requirements.txt
Django>=3.2.1,<3.3
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.22.1,<0.23
djangorestframework>=3.13.1,<3.14
Pillow>=9.3,<9.5
stripe>=2.60,<2.61